[ISSUE-199] i18n.js: bugfix. Support special chars
This commit arranges the issue #199 which was causing translation
issues.
Remote service call has been migrated to GET method, and separator has
been changed for a single non-common char.

The main issue behind the bug was the URL query-string scaping. Now the
query-string is built using the provided method by the request
library